apply some code review changes
- separate the DatabaseExtensions class from IDatabase
- rename ILazyLoading to ILazyLoader and use consistent naming for it
- separate the stuff in ILowlevelDatabaseIODevice

+++ tests/StringDB.Tests/StringDB10_0_0Tests.cs
@@ -272,7 +272,7 @@ namespace StringDB.Tests
 					var jumpPos = new byte[8];
 					ms.Read(jumpPos, 0, jumpPos.Length);
 
-					jumpPos.Should().BeEquivalentTo(BitConverter.GetBytes((long)0xDEADBEEF5));
+					jumpPos.Should().BeEquivalentTo(BitConverter.GetBytes(0xDEADBEEF5));
 				}
 			}
 		}
